+++ packages/expo-router/src/__tests__/getRoutes.test.web.ts
@@ -19,7 +19,6 @@ describe('rewrites', () => {
     ).toEqual({
       children: [
         {
-          children: [],
           contextKey: './old.js',
           destinationContextKey: './(app)/index.js',
           dynamic: null,
@@ -28,7 +27,6 @@ describe('rewrites', () => {
           type: 'rewrite',
         },
         {
-          children: [],
           contextKey: './(app)/index.js',
           dynamic: null,
           entryPoints: ['expo-router/build/views/Navigator.js', './(app)/index.js'],
@@ -65,7 +63,6 @@ describe('rewrites', () => {
     expect(routes).toEqual({
       children: [
         {
-          children: [],
           contextKey: './index.js',
           dynamic: null,
           entryPoints: ['expo-router/build/views/Navigator.js', './index.js'],
@@ -73,7 +70,6 @@ describe('rewrites', () => {
           type: 'route',
         },
         {
-          children: [],
           contextKey: './about.js',
           dynamic: null,
           entryPoints: ['expo-router/build/views/Navigator.js', './about.js'],
@@ -81,7 +77,6 @@ describe('rewrites', () => {
           type: 'route',
         },
         {
-          children: [],
           contextKey: './contact.js',
           dynamic: null,
           entryPoints: ['expo-router/build/views/Navigator.js', './contact.js'],
@@ -89,7 +84,6 @@ describe('rewrites', () => {
           type: 'route',
         },
         {
-          children: [],
           contextKey: './info.js',
           destinationContextKey: './about.js',
           dynamic: null,
@@ -98,7 +92,6 @@ describe('rewrites', () => {
           type: 'rewrite',
         },
         {
-          children: [],
           contextKey: './reach-us.js',
           destinationContextKey: './contact.js',
           dynamic: null,
@@ -134,7 +127,6 @@ describe('rewrites', () => {
     expect(routes).toEqual({
       children: [
         {
-          children: [],
           contextKey: './info.js',
           destinationContextKey: './(app)/index.js',
           dynamic: null,
@@ -143,7 +135,6 @@ describe('rewrites', () => {
           type: 'rewrite',
         },
         {
-          children: [],
           contextKey: './news.js',
           destinationContextKey: './(app)/index.js',
           dynamic: null,
@@ -152,7 +143,6 @@ describe('rewrites', () => {
           type: 'rewrite',
         },
         {
-          children: [],
           contextKey: './(app)/index.js',
           dynamic: null,
           entryPoints: ['expo-router/build/views/Navigator.js', './(app)/index.js'],
@@ -185,7 +175,6 @@ describe('rewrites', () => {
     ).toEqual({
       children: [
         {
-          children: [],
           contextKey: './(app)/index.js',
           dynamic: null,
           entryPoints: ['expo-router/build/views/Navigator.js', './(app)/index.js'],
@@ -193,7 +182,6 @@ describe('rewrites', () => {
           type: 'route',
         },
         {
-          children: [],
           contextKey: './(app)/[slug].js',
           dynamic: [
             {
@@ -206,7 +194,6 @@ describe('rewrites', () => {
           type: 'route',
         },
         {
-          children: [],
           contextKey: './old/[slug].js',
           destinationContextKey: './(app)/[slug].js',
           dynamic: [
@@ -246,7 +233,6 @@ describe('rewrites', () => {
     ).toEqual({
       children: [
         {
-          children: [],
           contextKey: './(app)/index.js',
           dynamic: null,
           entryPoints: ['expo-router/build/views/Navigator.js', './(app)/index.js'],
@@ -254,7 +240,6 @@ describe('rewrites', () => {
           type: 'route',
         },
         {
-          children: [],
           contextKey: './(app)/[slug].js',
           dynamic: [
             {
@@ -267,7 +252,6 @@ describe('rewrites', () => {
           type: 'route',
         },
         {
-          children: [],
           contextKey: './old/[slug].js',
           dynamic: [
             {
@@ -305,7 +289,6 @@ describe('rewrites', () => {
     ).toEqual({
       children: [
         {
-          children: [],
           contextKey: './+not-found.js',
           dynamic: [
             {
@@ -319,7 +302,6 @@ describe('rewrites', () => {
           type: 'route',
         },
         {
-          children: [],
           contextKey: './404.js',
           destinationContextKey: './+not-found.js',
           dynamic: null,
@@ -355,7 +337,6 @@ describe('middleware', () => {
     ).toEqual({
       children: [
         {
-          children: [],
           contextKey: './index.js',
           dynamic: null,
           entryPoints: ['expo-router/build/views/Navigator.js', './index.js'],